requirements.txt
requests
beautifulsoup4
python-dotenv
selenium
webdriver-manager
openai==0.28.1
